ORC-418. Fix broken centos6 compilation.
Fixes #328

+++ c++/src/RLEv1.hh
@@ -40,7 +40,7 @@ public:
 private:
     int64_t delta;
     bool repeat;
-    int tailRunLength;
+    uint64_t tailRunLength;
 
     void write(int64_t val) override;
     void writeValues();
